pyproject.toml
[tool.poetry]
name = "gem"
version = "0.1.0"
description = ""
authors = ["Boris Zhenikhov <zhenikhov@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "*"
pyyaml = "^6.0.1"
jsonschema = "^4.23.0"
ruamel-yaml = "^0.18.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
